[Bronze V] Title: 이상한 기호, Time: 176 ms, Memory: 17716 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -3,10 +3,10 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
-        int n1 = scan.nextInt();
-        int n2 = scan.nextInt();
+        long n1 = scan.nextInt();
+        long n2 = scan.nextInt();
         
-        int result = (n1 + n2) * (n1 - n2);
+        long result = (n1 + n2) * (n1 - n2);
         System.out.println(result);
     }
 }
